# swagger-docs.yml
openapi: "3.0.0"
info:
  title: "TypeScript Node CRUD Docs"
  version: "1.0.0"
  description: "TypeScript based node js basic CRUD apis with MongoDB"
  license:
    name: "MIT"
servers:
  - url: "http://localhost:5003/api/v1"
  - url: "https://uat.myserver.com/api/v1"
...
paths:
...
  /user/remove:
    delete:
      tags:
        - "User"
      summary: "delete a user"
      description: "You can delete a user with this api via object id."
      operationId: "removeUser"
      parameters:
        - name: "id"
          required: true,
          in: "body"
          content: 
            application/x-www-form-urlencoded: {}
          schema: 
            type: "string"
            example:
              id: "XXXXXXXXXXXXXXXXXXXXXXXX"
      requestBody:
        content:
          application/json:
            schema:
              type: "object" # raw object
              example:
                id: "XXXXXXXXXXXXXXXXXXXXXXXX"
      responses:
        200:
          description: "Successful response"
          content:
            application/json: {}
        400:
          description: "Bad Request"
          content:
            application/json: {}
        500:
          description: "Internal Server Error [Learn More](myErrorDescription)"
          content:
            application/json: {}
          links:
            Documentation:
              description: "Please check more details regarding error on URL"
              url: "https://example.com/create-user-docs"
...